chore: ファイル名やダミー画像の整理
