+++ src/assets/images/battle/action/battle-debuff-sample.png
