+++ src/components/Battle/Partials/BattleBuffIcon.module.scss
@@ -5,5 +5,5 @@
   border: 2px solid var(--color-beige-darken);
   backdrop-filter: blur(4px) sepia(0.15);
   filter: drop-shadow(1px 1px var(--color-darkblue));
-  padding: 0.5rem;
+  padding: 0.25rem;
 }
