+++ src/components/battle/enemy/BattleEnemyHpBar.module.scss
@@ -1,6 +1,6 @@
 .enemy-hp-bar {
   background-color: var(--color-darkblue);
-  height: 0.6rem;
+  height: 0.5rem;
   margin: 0 20vw;
   position: relative;
   width: 60vw;
@@ -27,6 +27,6 @@
     position: absolute;
     top: -2rem;
     width: 4rem;
-    filter: drop-shadow(4px 4px var(--color-darkblue));
+    filter: drop-shadow(3px 3px var(--color-darkblue));
   }
 }
