+++ src/interfaces/battle.d.ts
