+++ src/styles/globals.scss
@@ -61,6 +61,7 @@ img {
 @import url("https://fonts.googleapis.com/css2?family=New+Tegomin&family=Noto+Sans+JP:wght@100&family=Shippori+Mincho:wght@500&display=swap");
 
 body {
+  min-height: 100vh;
   margin: 0;
   padding: 0;
   background-color: var(--color-white);
